requirements.txt
requests>=2.31
pandas>=1.5
Jinja2>=3.1
python-dotenv>=1.0
supabase>=1.0.0
stripe>=8.0
